16t everywhere

+++ src/software/cnn_io.c
@@ -24,7 +24,7 @@ struct my_comp {
     int value;
 } dev;
 
-static void write_value(int val[], int max_addr){
+static void write_value(int16_t val[], int max_addr){
     /* iowrite8(value, adress-to-write-to)*/
     //int addr = 0;
     //int[] val = *val_addr
@@ -46,10 +46,10 @@ static void write_value(int val[], int max_addr){
 };
 
 // cannot return array so will return pointer to array
-static int* read_value(int addr, int max_addr){
+static int16_t* read_value(int addr, int max_addr){
     /* ioread(adress-to-read-from)*/
     //static int out[max_addr-addr]; // doesnt work because dynamic size and static (needs static to retain mem addr outside the fucntion)
-    int* out_ptr = kmalloc(sizeof(int)*(max_addr-addr), GFP_KERNEL); // dynamic allocation
+    int16_t* out_ptr = kmalloc(sizeof(int16_t)*(max_addr-addr), GFP_KERNEL); // dynamic allocation
     int addr_local;
     for (addr_local = 0; addr_local < max_addr-addr; addr_local = addr_local + 1){
         *(out_ptr+addr_local) = ioread16(dev.virtbase+addr+addr_local);
@@ -63,10 +63,10 @@ static long cnn_ioctl(struct file *f, unsigned int cmd, unsigned long val_arg)
 {
     // new array of same size as input
     // changes
-    int (*arr_ptr)[10] = val_arg; // int (*arr_ptr)[10] = val_arg;
+    int16_t (*arr_ptr)[10] = val_arg; // int (*arr_ptr)[10] = val_arg;
     //int (*a)[10] = l;
     pr_info("iooctl: val_local size %d", sizeof(*arr_ptr)/sizeof((*arr_ptr)[0]));
-    int val_local[sizeof(*arr_ptr)/sizeof((*arr_ptr)[0])];
+    int16_t val_local[sizeof(*arr_ptr)/sizeof((*arr_ptr)[0])];
 
     switch(cmd){
         case CNN_WRITE_VAL:;
@@ -91,7 +91,7 @@ static long cnn_ioctl(struct file *f, unsigned int cmd, unsigned long val_arg)
             //if ((val_local = read_value()) != 0) 
             //    return -EACCES;
             //int *arr_ptr_local;
-            int* arr_ptr_local = read_value(0,10);
+            int16_t* arr_ptr_local = read_value(0,10);
             pr_info("ictl_reading: done reading %d", arr_ptr_local);
             pr_info("ictl_reading: val_local[0] %d", *(arr_ptr_local));
             pr_info("\n");
